+++ Assets/VRM10/Runtime/IO/Vrm10Storage.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
 using System.Runtime.InteropServices;
+using UniJSON;
 using VrmLib;
 
 
@@ -478,6 +479,18 @@ namespace UniVRM10
         {
             int index = Gltf.buffers.IndexOf(buffer);
             return Buffers[index].Bytes;
+        }
+
+        public byte[] ToBytes()
+        {
+            Gltf.buffers[0].byteLength = Buffers[0].Bytes.Count;
+
+            var f = new JsonFormatter();
+            UniGLTF.GltfSerializer.Serialize(f, Gltf);
+            var json = f.GetStoreBytes();
+
+            var glb = UniGLTF.Glb.Create(json, Buffers[0].Bytes);
+            return glb.ToBytes();
         }
     }
 }
